+++ doc/release-notes.md
@@ -4,3 +4,9 @@ release-notes at release time)
 Notable changes
 ===============
 
+RPC Changes
+-----------
+
+* The RPC methods `getnetworkhashps`, `keypoolrefill`, `settxfee`,
+  `createrawtransaction`, `fundrawtransaction`, and `signrawtransaction`
+  have been deprecated, but are still enabled by default.